refactor: extract board scanning adapters
